Solamente presentar fichas sin asignar, abstracción de los Queries y Mutations para inspección
